% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getActualGeneration}
\alias{getActualGeneration}
\title{getActualGeneration}
\usage{
getActualGeneration(startDate, endDate)
}
\arguments{
\item{startDate}{date format is YYYY-MM-DD Distinguishing of hours or minutes is not allowed. If you e.g. only need half of a day, please subset the data yourself.}

\item{endDate}{date format is YYYY-MM-DD Distinguishing of hours or minutes is not allowed. If you e.g. only need half of a day, please subset the data yourself.}
}
\value{
a data.frame object with the actual power generation per product type of the four german TSO
}
\description{
This method retrieves the actual energy generation for different product typse of the four german TSO and the Netzregelverbund (summation).
The resolution of the data is in discrete 15 minutes timesteps (CET (UTC+1) / CEST (UTC+2)) and the values are presented in MW.
The product types are Biomass, Fossil Brown coal/Lignite, Fossil Coal-derived gas, Fossil Gas, Fossil Hard coal, Fossil Oil, Fossil Oil shale, Fossil Peat, Geothermal, Hydro Pumped Storage, Hydro Run-of-river and poundage, Hydro Water Reservoir, Marine, Nuclear, Other, Other renewable, Solar, Waste, Wind Offshore, Wind Onshore.

NOTE:
 - Always downloads for the whole year and then subsets the data to the given time period (3-4MB). This is done for each TSO
 - For every request a Login POST request is needed. For Login a pseudo account is used.
}
\examples{
loadData <- getActualGeneration("2017-01-01", "2017-06-30")

}
