% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapeData.R
\name{getWindSolarDayAheadGeneration}
\alias{getWindSolarDayAheadGeneration}
\title{getWindSolarDayAheadGeneration}
\usage{
getWindSolarDayAheadGeneration(startDate, endDate)
}
\arguments{
\item{startDate}{date format is YYYY-MM-DD Distinguishing of hours or minutes is not allowed. If you e.g. only need half of a day, please subset the data yourself.}

\item{endDate}{date format is YYYY-MM-DD Distinguishing of hours or minutes is not allowed. If you e.g. only need half of a day, please subset the data yourself.}
}
\value{
a data.frame object with the day-ahead generation forecast in Wind (Onshore, Offshore) and Solar of the four german TSO
}
\description{
This method retrieves the actual and one day-ahead forecast of the energy generation in Wind (Onshore, Offshore) and Solar for the four german TSO and the Netzregelverbund (summation).
The resolution of the data is in discrete 15 minutes timesteps (CET (UTC+1) / CEST (UTC+2)) and the values are presented in MW.

NOTE:
 - Always downloads for the whole year and then subsets the data to the given time period (1-2MB). This is done for each TSO
 - For every request a Login POST request is needed. For Login a pseudo account is used.
}
\examples{
loadData <- getWindSolarDayAheadVsActual("2017-01-01", "2017-06-30")

}
